refactor(core): extract i18n tests to sibling file
Move the single inline `#[cfg(test)] mod tests` from
primer-core/src/i18n.rs into a sibling file src/i18n/tests.rs.
Pure verbatim test-module relocation: the test module is a
descendant, so `use super::*` resolves to `i18n` unchanged — no
visibility edits, no behaviour change.

i18n.rs: 539 -> 252 lines (-287). All 176 primer-core lib tests
preserved; workspace 51 ok blocks, 0 failures; clippy + fmt clean.
